+++ examples/advanced/src/gen/clients/swr/userController/useGetUserByNameSWR.ts
@@ -35,9 +35,10 @@ export function useGetUserByNameSWR(
     query?: Parameters<typeof useSWR<ResponseConfig<GetUserByNameQueryResponse>, ResponseErrorConfig<GetUserByName400 | GetUserByName404>>>[2]
     client?: Partial<RequestConfig> & { client?: typeof fetch }
     shouldFetch?: boolean
+    immutable?: boolean
   } = {},
 ) {
-  const { query: queryOptions, client: config = {}, shouldFetch = true } = options ?? {}
+  const { query: queryOptions, client: config = {}, shouldFetch = true, immutable } = options ?? {}
 
   const queryKey = getUserByNameQueryKeySWR({ username })
 
@@ -45,6 +46,13 @@ export function useGetUserByNameSWR(
     shouldFetch ? queryKey : null,
     {
       ...getUserByNameQueryOptionsSWR({ username }, config),
+      ...(immutable
+        ? {
+            revalidateIfStale: false,
+            revalidateOnFocus: false,
+            revalidateOnReconnect: false,
+          }
+        : {}),
       ...queryOptions,
     },
   )